+++ Scripts/Install-Modules.ps1
@@ -2,7 +2,7 @@
 $PackageProvider = Install-PackageProvider -Name "NuGet" -Force
 
 # Install required modules
-$Modules = @("Evergreen", "IntuneWin32App", "Az.Storage", "Az.Resources")
+$Modules = @("Evergreen", "IntuneWin32App", "MSGraphRequest", "Az.Storage", "Az.Resources")
 foreach ($Module in $Modules) {
     Write-Output -InputObject "Attempting to install the following module: $($Module)"
     Install-Module -Name $Module -Force -Confirm:$false